Refactor auth code into separate file
