+++ fireblog/__init__.py
@@ -1,8 +1,5 @@
 from pyramid.config import Configurator
 from sqlalchemy import engine_from_config
-from sqlalchemy.orm.exc import NoResultFound
-from pyramid.security import Allow, ALL_PERMISSIONS
-from pyramid.authentication import AuthTktAuthenticationPolicy
 from pyramid.events import BeforeRender
 from pyramid.response import Response
 import fireblog.utils as utils
@@ -68,45 +65,6 @@ def add_urlify_function(event):
 
 def add_settings_dict_to_templates(event):
     event['settings_dict'] = settings_dict
-
-
-def groupfinder(userid, request) -> list:
-    """Looks up and returns the groups the userid belongs to.
-    If the userid doesn't exist, they are created as a commenter, and the
-    group they belong to (g:commenter) is returned."""
-    query = DBSession.query(Users). \
-        filter(Users.userid == userid)
-    try:
-        user = query.one()
-        return [user.group]
-    except NoResultFound:
-        group = create_commenter_and_return_group(userid)
-        return [group]
-
-
-def create_commenter_and_return_group(userid) -> str:
-    """This function assumes userid doesn't exist in the db, and creates a new
-    user with this userid, as a commenter.
-
-    :return: group the user belongs to (g:commenter)"""
-    group = 'g:commenter'
-    new_user = Users(userid=userid, group=group)
-    DBSession.add(new_user)
-    log.info('New commenter {} has been created'.format(userid))
-    return group
-
-
-class Root(object):
-    """Resource tree to map groups to permissions. We allow admins to do
-    anything, and commenters to be able to comment only.
-    """
-    __acl__ = [
-        (Allow, 'g:admin', ALL_PERMISSIONS),
-        (Allow, 'g:commenter', 'add-comment'),
-    ]
-
-    def __init__(self, request):
-        self.request = request
 
 
 def add_routes(config):
@@ -181,19 +139,13 @@ def main(global_config, **settings):
         if not name.startswith('fireblog'):
             settings[name] = value
 
-    config = Configurator(settings=settings, root_factory=Root)
+    config = Configurator(settings=settings)
     config.include('pyramid_mako')
-    config.include("pyramid_persona")
+    config.include('fireblog.login')
     config.add_static_view(name='bower', path='fireblog:../bower_components')
     config.add_request_method(get_bower_url)
     config.add_response_adapter(
         template_response_adapter, utils.TemplateResponseDict)
-    authn_policy = AuthTktAuthenticationPolicy(
-        settings['persona.secret'],
-        callback=groupfinder)
-    config.set_authentication_policy(authn_policy)
-    # Pyramid_persona has already set an authorization policy, so
-    # this has not been done here.
     include_all_components(config)
     config.scan()
     return config.make_wsgi_app()
